Create task 13
Problem 1
Not using class string from the Standard Library (header string), write your own
class String, objects of which represents strings (arrays of characters) of arbitrary
length. You may use functions from header cstring like strlen, strcpy, strcat or
strcmp and also functions operating on single characters (like toupper or tolower)
from header cctype.
Dene appropriate constructors and destructor. Dene also methods and friend functions, in particular overloading operators of addition (concatenation) of strings, of
comparisons and assignment. They should also work when one of the arguments is
a plain C-string.
Scheme of the program







Methods toLower and toUpper (changing the case of characters in a string) modify
and return by reference the object they were invoked on. Method length returns the
number of characters in the string represented by the object (not counting the '\0'
character at the end).
The program above, after implementing all functions, should print
To be or not to be
TO BE OR NOT TO BE!
true